requirements.txt
asgiref==3.2.10
Django==3.0
django-phone-field==1.8.1
et-xmlfile==1.0.1
future==0.18.2
jdcal==1.4.1
numpy==1.19.1
openpyxl==3.0.4
pandas==1.1.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
xlrd==1.2.0
